requirements.txt
cdsapi
pyyaml
torch
pandas
numpy
matplotlib
scikit-learn
xarray
joblib